requirements.txt
disnake
aiohttp
asyncio
aiofiles
psutil
python-dateutil
munch
motor
bs4